refactor(LottoRule):로또규칙을 config 폴더에 넣는게 알맞는 것 같아 폴더 변경
