+++ src/test/java/lotto/model/WinningNumberTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import lotto.config.LottoRule;
 import lotto.validation.ErrorMessage;
-import lotto.validation.LottoRule;
 
 public class WinningNumberTest {
